% Auto-generated: do not edit by hand
\name{''SimulationComponent}

\alias{''SimulationComponent}

\title{SimulationComponent component}

\description{

}

\usage{
''SimulationComponent(id=NULL, tick_data=NULL)
}

\arguments{
\item{id}{Character. Unique ID to identify this component in Dash callbacks.}

\item{tick_data}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

